chore(tests): split express/restify tests
