+++ package.json
@@ -49,7 +49,8 @@
     "pretest": "standard",
     "test": "mocha -R spec ./test/unit.js && mocha -R spec ./test/integration/resource_filter.js --timeout 10s && mocha -R spec ./test/integration/express.js --timeout 10s && mocha -R spec ./test/integration/restify.js --timeout 10s",
     "coverage-unit": "istanbul cover node_modules/mocha/bin/_mocha -- -R spec ./test/unit.js",
-    "coverage-integration": "istanbul cover node_modules/mocha/bin/_mocha -- -R spec ./test/integration.js --timeout 10s"
+    "coverage-express": "istanbul cover node_modules/mocha/bin/_mocha -- -R spec ./test/express.js --timeout 10s",
+    "coverage-restify": "istanbul cover node_modules/mocha/bin/_mocha -- -R spec ./test/restify.js --timeout 10s"
   },
   "standard": {
     "globals": [
